Modified the framework to consistently keep track of the agent statistics.
Added a variable "ncond" to keep track of the number of conditions found in
a dataset.

Added vector<> structures to the agent to keep track of the mean, standard
deviation, and the post-threshold-count of the entries for each condition.
Added a calcStats() method to calculate statistics for all conditions for
a given agent. Also added a showStats() method to print agent data in stdout.
The computed statistics for each condition are of the threshold-ed data
based on the "istate" variable of the entry object.

Added loops to the dataset methods that handle and use the agent objects
such that the agents are always kep up-to-date with any data thresholding.
At present, on-demand re-creation of the agents will calculate proper stats.

Restructured the driver function and modified the Makefile.
IN 2017/05/07